requirements.txt
clickhouse_driver==0.2.5
instaloader==4.9.6
moviepy==1.0.3
mysql-connector-python==8.0.23
Pillow==9.4.0
pytesseract==0.3.10
requests==2.28.2
yadisk==1.3.1
